# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wishlist:
    get:
      operationId: listWishlist
      tags: [wishlist]
      summary: List wishlist items (auth required)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistItem"
    post:
      operationId: addWishlistItem
      tags: [wishlist]
      summary: Add product to wishlist (auth required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistItem"
  /wishlist/{productId}:
    delete:
      operationId: removeWishlistItem
      tags: [wishlist]
      summary: Remove product from wishlist (auth required)
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Removed
...
components:
...
  schemas:
...
    ProductSummary:
      type: object
      required:
        [
          id,
          name,
          slug,
          price,
          salePrice,
          images,
          rating,
          reviewCount,
          stock,
        ]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        price: { type: number }
        salePrice: { type: ["number", "null"] }
        images: { type: array, items: { type: string } }
        rating: { type: number }
        reviewCount: { type: integer }
        stock: { type: integer }
        featured: { type: boolean }
        trending: { type: boolean }
        bestSeller: { type: boolean }
        newArrival: { type: boolean }
...
    WishlistItem:
      type: object
      required: [id, productId, product, createdAt]
      properties:
        id: { type: integer }
        productId: { type: integer }
        product:
          $ref: "#/components/schemas/ProductSummary"
        createdAt: { type: string, format: date-time }
    WishlistItemInput:
      type: object
      required: [productId]
      properties:
        productId: { type: integer }
...